I regret making things uint

+++ OpenDreamRuntime/ByondApi/ByondApi.Functions.cs
@@ -5,6 +5,7 @@ using OpenDreamRuntime.Objects;
 using OpenDreamRuntime.Objects.Types;
 using OpenDreamRuntime.Procs;
 using OpenDreamRuntime.Procs.Native;
+using OpenDreamShared.EngineUtils;
 
 // ReSharper disable InconsistentNaming
 
@@ -800,7 +801,7 @@ public static unsafe partial class ByondApi {
 
         return RunOnMainThread<byte>(() => {
             try {
-                if (_dreamMapManager!.TryGetTurfAt(new Vector2i(xyz->x, xyz->y), xyz->z, out var turf)) {
+                if (_dreamMapManager!.TryGetTurfAt(new Vector2i(xyz->x, xyz->y), (uint)xyz->z, out var turf)) {
                     DreamValue val = new(turf);
                     *result = ValueToByondApi(val);
                 } else {
